add class, reviews and more
